user

+++ app/Http/Controllers/TasksController.php
@@ -11,58 +11,69 @@ class TasksController extends Controller
     
     public function index()
     {
-        $tasks = Task::all();
-        return view('tasks.index',['tasks'=>$tasks,]);
-    }
-
-   
-    public function create()
-    {
-        $task = new Task;
- 	    return view('tasks.create',['task'=>$task,]);
+        $data=[];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = ['user' =>$user,'tasks' =>$tasks];
+            $data += $this->counts($user);
+            return view('users.show',$data);
+        }else{
+            return view('welcome');
+        }
     }
 
     
     public function store(Request $request)
     {
         $this->validate($request, ['status' => 'required','content' => 'required|max:191',]);
-        $task = new Task;
-        $task->status=$request->status;
-        $task->content=$request->content;
-        $task->save();
-        return redirect ('/');
+        
+        $task->user()->tasks()->create(['status' => $request->status,]);
+        $task->user()->tasks()->create(['content' => $request->content,]);
+        
+        return redirect()->back();
     }
 
     
     public function show($id)
     {
-        $task = Task::find($id);
-        return view('tasks.show',['task'=>$task,]);
+        $task = \App\Task::find($id);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.show',['task'=>$task,]);
+        }
     }
 
     
     public function edit($id)
     {
-        $task = Task::find($id);
-        return view('tasks.edit',['task'=>$task,]);
+        if (\Auth::id() === $task->user_id){
+            return view('tasks.edit',['task'=>$task,]);
+        }
     }
 
     
     public function update(Request $request, $id)
     {
+        $task = \App\Task::find($id);
+        if (\Auth::id() === $task->user_id) {
         $this->validate($request, ['status' => 'required','content' => 'required|max:191',]);
-        $task = Task::find($id);
-        $task->status=$request->status;
-        $task->content=$request->content;
-        $task->save();
-        return redirect('/');
+        
+        $task->user()->tasks()->create(['status' => $request->status,]);
+        $task->user()->tasks()->create(['content' => $request->content,]);
+        }
+        return redirect()->back();
     }
 
     
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
-        return redirect('/');
+        $task = \App\Task::find($id);
+
+        if (\Auth::id() === $task->user_id){
+            $task->delete();
+        }
+            return redirect()->back();
+        
     }
 }
